+++ custom_components/battery_controller/battery_model.py
@@ -332,100 +332,6 @@ def calculate_new_soc(
     return new_soc, actual_energy
 
 
-def calculate_max_charge_power(
-    current_soc_kwh: float,
-    duration_hours: float,
-    config: BatteryConfig,
-) -> float:
-    """Calculate maximum charge power considering SoC limits.
-
-    Args:
-        current_soc_kwh: Current state of charge in kWh
-        duration_hours: Duration in hours
-        config: Battery configuration
-
-    Returns:
-        Maximum charge power in kW
-    """
-    # Energy needed to reach max SoC
-    energy_headroom = config.max_soc_kwh - current_soc_kwh
-
-    if energy_headroom <= 0 or duration_hours <= 0:
-        return 0.0
-
-    # Power needed to fill in duration (before efficiency)
-    current_soc_percent = (current_soc_kwh / config.capacity_kwh) * 100.0
-    efficiency = calculate_efficiency(
-        config.max_charge_power_kw, current_soc_percent, config
-    )
-
-    power_for_headroom = energy_headroom / (duration_hours * efficiency)
-
-    return min(power_for_headroom, config.max_charge_power_kw)
-
-
-def calculate_max_discharge_power(
-    current_soc_kwh: float,
-    duration_hours: float,
-    config: BatteryConfig,
-) -> float:
-    """Calculate maximum discharge power considering SoC limits.
-
-    Args:
-        current_soc_kwh: Current state of charge in kWh
-        duration_hours: Duration in hours
-        config: Battery configuration
-
-    Returns:
-        Maximum discharge power in kW (as positive value)
-    """
-    # Energy available above min SoC
-    energy_available = current_soc_kwh - config.min_soc_kwh
-
-    if energy_available <= 0 or duration_hours <= 0:
-        return 0.0
-
-    # Power needed to drain in duration (after efficiency)
-    current_soc_percent = (current_soc_kwh / config.capacity_kwh) * 100.0
-    efficiency = calculate_efficiency(
-        -config.max_discharge_power_kw, current_soc_percent, config
-    )
-
-    power_for_available = energy_available * efficiency / duration_hours
-
-    return min(power_for_available, config.max_discharge_power_kw)
-
-
-def should_cycle(
-    buy_price: float,
-    sell_price: float,
-    rte: float,
-    degradation_per_kwh: float,
-) -> bool:
-    """Check if cycling the battery is profitable.
-
-    Only cycle if profitable after RTE losses and degradation costs.
-
-    Minimum spread needed:
-    sell_price > buy_price / rte + degradation_per_kwh * 2
-
-    Example with RTE=0.90, degradation=0.03:
-    sell_price > buy_price / 0.90 + 0.06
-    At buy_price = 0.10: sell_price > 0.17 needed
-
-    Args:
-        buy_price: Price to buy electricity (EUR/kWh)
-        sell_price: Price to sell electricity (EUR/kWh)
-        rte: Round trip efficiency (0-1)
-        degradation_per_kwh: Degradation cost per kWh throughput
-
-    Returns:
-        True if cycling is profitable
-    """
-    min_sell_price = buy_price / rte + degradation_per_kwh * 2
-    return sell_price > min_sell_price
-
-
 def calculate_degradation_cost_per_kwh(
     replacement_cost_per_kwh: float = 500.0,
     lifecycle_cycles: int = 6000,
